backend/requirements.txt
# NovaSight Backend Dependencies
# ===============================
# Python packages for Flask API

# Flask Core
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0

# Database
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Validation
pydantic==2.5.3
email-validator==2.1.0

# HTTP Client
httpx==0.26.0

# Redis
redis==5.0.1

# Encryption
cryptography==41.0.7

# Template Engine
Jinja2==3.1.2

# WSGI Server
gunicorn==21.2.0

# ClickHouse
clickhouse-connect==0.7.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# Logging
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-flask==1.3.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==22.0.0

# Development
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
